+++ lib/opam.mli
@@ -14,8 +14,7 @@ end
 
 module Package_summary : sig
   type t = {
-    name : string;
-    version : string;
+    package : OpamPackage.t;
     url_src : Url.t option;
     hashes : OpamHash.t list;
     dev_repo : string option;
@@ -41,7 +40,11 @@ end
 module Pp : sig
   val package : OpamPackage.t Fmt.t
 
+  val raw_package : OpamPackage.t Fmt.t
+
   val package_name : OpamPackage.Name.t Fmt.t
+
+  val version : OpamPackage.Version.t Fmt.t
 
   val hash : OpamHash.t Fmt.t